Py files to run crnn models.
Run 'python3 main.py' to run the crnn model. Need to upload segments and csv files containing information of segments to hpc server first. If interested in the detailed information of the modeling process, it is better to run the code in 'main.py' line by line and make any changes if necessary.